% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerceVector.R
\name{coerceVector}
\alias{coerceVector}
\alias{Rf_coerceVector}
\alias{AS_LOGICAL}
\alias{AS_INTEGER}
\alias{AS_NUMERIC}
\alias{AS_CHARACTER}
\alias{AS_COMPLEX}
\alias{AS_VECTOR}
\alias{AS_LIST}
\alias{AS_RAW}
\title{Coerces a Vector}
\source{
Declaration: \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{src/include/Rinternals.h} and \href{https://github.com/wch/r-source/blob/trunk/src/include/Rdefines.h}{src/include/Rdefines.h}

Implementation: \href{https://github.com/wch/r-source/blob/trunk/src/main/coerce.c}{src/main/coerce.c}
}
\usage{
coerceVector(v, type)
}
\arguments{
\item{v}{(\link{SEXP}) An \R object.}

\item{type}{(\link{SEXPTYPE}) An macro constant of length one.}
}
\value{
\code{coerceVector()} returns (\link{SEXP}) of type \link{SEXPTYPE} \code{type}.
}
\description{
Coerces a Vector
}
\section{C API}{
\if{html}{\out{<div class="c">}}\preformatted{#include <Rinternals.h>
SEXP Rf_coerceVector(SEXP, SEXPTYPE)
#define coerceVector		Rf_coerceVector
}\if{html}{\out{</div>}}

Related macros:\if{html}{\out{<div class="c">}}\preformatted{#include <Rdefines.h>
#define AS_LOGICAL(x)		Rf_coerceVector(x,LGLSXP)
#define AS_INTEGER(x)		Rf_coerceVector(x,INTSXP)
#define AS_NUMERIC(x)		Rf_coerceVector(x,REALSXP)
#define AS_CHARACTER(x)		Rf_coerceVector(x,STRSXP)
#define AS_COMPLEX(x)		Rf_coerceVector(x,CPLXSXP)
#define AS_VECTOR(x)		Rf_coerceVector(x,VECSXP)
#define AS_LIST(x)		Rf_coerceVector(x,VECSXP)
#define AS_RAW(x)		Rf_coerceVector(x,RAWSXP)
}\if{html}{\out{</div>}}
}

\section{R API}{
\if{html}{\out{<div class="r">}}\preformatted{library(base)

x <- 1:5
y_integer   <- as.integer(x)
y_logical   <- as.logical(x)
y_numeric   <- as.numeric(x)
y_character <- as.character(x)
}\if{html}{\out{</div>}}
}

