% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAttrib.R
\name{getAttrib}
\alias{getAttrib}
\alias{setAttrib}
\title{Gets and Sets an Attribute of an R Object}
\source{
Declaration: \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{src/include/Rinternals.h}

Implementation: \href{https://github.com/wch/r-source/blob/trunk/src/main/attrib.c}{src/main/attrib.c}
}
\usage{
getAttrib(vec, name)

setAttrib(vec, name, val)
}
\arguments{
\item{vec}{(\link{SEXP}) An \R object.}

\item{name}{(\link{SEXP}) A \link{character} of length one.}

\item{val}{(\link{SEXP}) An \R object.}
}
\value{
\code{getAttrib()} returns (\link{SEXP}) the value of attribute \code{name}.

\code{setAttrib()} returns (\link{SEXP}) a copy of \R object \code{vec} with attribute \code{name} set.

(\link{SEXP})
}
\description{
Gets and Sets an Attribute of an R Object
}
\section{C API}{


\if{html}{\out{<div class="sourceCode c">}}\preformatted{#include <Rinternals.h>

SEXP getAttrib(SEXP vec, SEXP name)
SEXP setAttrib(SEXP vec, SEXP name, SEXP val)
}\if{html}{\out{</div>}}
}

\section{R API}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(base)

value <- attr(vec, name, exact = TRUE)
attr(vec, name) <- val
}\if{html}{\out{</div>}}
}

