% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isNumeric.R
\name{isNumeric}
\alias{isNumeric}
\alias{Rf_isNumeric}
\title{Checks if an SEXP is of Certain Type}
\source{
Declaration: \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{src/include/Rinternals.h} and \href{https://github.com/wch/r-source/blob/trunk/src/include/Rdefines.h}{src/include/Rdefines.h}

Implementation: Inline function \verb{isNumeric(SEXP s)}, cf. \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinlinedfuns.h}{src/include/Rinlinedfuns.h}
}
\usage{
isNumeric(s)
}
\arguments{
\item{s}{(\link{SEXP}) An \R object.}
}
\value{
(\code{Rboolean}) \code{true} (=1) or \code{false} (=0).
}
\description{
Checks if an SEXP is of Certain Type
}
\section{C API}{
\if{html}{\out{<div class="c">}}\preformatted{#include <Rinternals.h>
Rboolean Rf_isNumeric(SEXP);
#define isNumeric		Rf_isNumeric
}\if{html}{\out{</div>}}
}

\section{R API}{
\if{html}{\out{<div class="r">}}\preformatted{library(base)

is.numeric(integer(length = 2L))  ## TRUE
is.numeric(numeric(length = 2L))  ## TRUE
is.numeric(factor(c("a", "b")))   ## FALSE
is.numeric(logical(length = 2L))  ## FALSE but isNumeric() = true!
}\if{html}{\out{</div>}}
}

