% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_CheckUserInterrupt.R
\name{R_CheckUserInterrupt}
\alias{R_CheckUserInterrupt}
\title{Checks for User Interrupt}
\source{
Declaration: \href{https://github.com/wch/r-source/blob/trunk/src/include/R_ext/Utils.h}{src/include/R_ext/Utils.h}

Implementation: \href{https://github.com/wch/r-source/blob/trunk/src/main/errors.c}{src/main/errors.c}
}
\usage{
R_CheckUserInterrupt(void)
}
\arguments{
\item{void}{(void) Nothing.}
}
\value{
Nothing.
}
\description{
Check if the user has requested an interrupt, and if so branch to
\R's error signaling functions.
}
\section{C API}{
\if{html}{\out{<div class="c">}}\preformatted{#include <R_ext/Utils.h>

void R_CheckUserInterrupt(void)
}\if{html}{\out{</div>}}
}

\section{R API}{


All evaluation at the top level of R checks for interrupts.

Related functions that affect \code{R_CheckUserInterrupt()}:
\itemize{
\item \code{\link[base:suspendInterrupts]{base::suspendInterrupts()}} - suspends interrupts during evaluation
\item \code{\link[base:allowInterrupts]{base::allowInterrupts()}} - allows interrupts during evaluation
}
}

