% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEXPTYPE.R
\docType{class}
\name{SEXPTYPE}
\alias{SEXPTYPE}
\title{The \code{SEXPTYPE} Structure}
\format{
A C struct.
}
\usage{
SEXPTYPE
}
\description{
The \code{SEXPTYPE} Structure
}
\section{C API}{
\if{html}{\out{<div class="c">}}\preformatted{#include <Rinternals.h>
}\if{html}{\out{</div>}}
}

\section{Common types}{


Common \code{SEXPTYPE}:s an their \R equivalent, if available:
\itemize{
\item \code{REALSXP}	\link{numeric} with storage mode \link{double}
\item \code{INTSXP}	\link{integer}
\item \code{CPLXSXP}	\link{complex}
\item \code{LGLSXP}	\link{logical}
\item \code{STRSXP}	\link{character}
\item \code{VECSXP}	list (generic vector)
\item \code{LISTSXP}	\link{pairlist}
\item \code{DOTSXP}	a \link{...} object
\item \code{NILSXP}	\link{NULL}
\item \code{SYMSXP}	name/symbol
\item \code{CLOSXP}	\link{function} or function closure
\item \code{ENVSXP}	\link{environment}
}
}

\seealso{
The \code{SEXPTYPE} is documented in Section 'Details of R types' of 'R Language Definition'.
}
\keyword{datasets}
