% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAttrib.R
\name{getAttrib}
\alias{getAttrib}
\alias{setAttrib}
\title{Gets and Sets an Attribute of an R Object}
\source{
\href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{src/include/Rinternals.h}

\href{https://github.com/wch/r-source/blob/trunk/src/main/attrib.c}{src/main/attrib.c}
}
\usage{
getAttrib(vec, name)

setAttrib(vec, name, val)
}
\arguments{
\item{vec}{(\link{SEXP}) An \R object.}

\item{name}{(\link{SEXP}) A \link{character} of length one.}

\item{val}{(\link{SEXP}) An \R object.}
}
\value{
(\link{SEXP}) Gets the value of attribute \code{name}.

(\link{SEXP}) A copy of \R object \code{SEXP_vec} with attribute \code{name} set.
}
\description{
Gets and Sets an Attribute of an R Object
}
\section{C API}{
\preformatted{SEXP getAttrib(SEXP vec, SEXP name)
SEXP setAttrib(SEXP vec, SEXP name, SEXP val)
}
}

\section{R API}{
\preformatted{value <- attr(vec, name)
attr(vec, name) <- val
}
}

