% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEXP.R
\docType{class}
\name{SEXP}
\alias{SEXP}
\title{The \code{SEXP} Structure}
\format{
A C struct.
}
\usage{
SEXP
}
\description{
The \code{SEXP} Structure
}
\section{C API}{


\if{html}{\out{<div class="sourceCode c">}}\preformatted{#include <Rinternals.h>
}\if{html}{\out{</div>}}
}

\seealso{
The \code{SEXP} structure is documented in 'R Language Definition'.
}
\keyword{datasets}
